requirements.txt
livekit-agents~=1.3.10
livekit-plugins-openai~=1.3.9
livekit-plugins-cartesia~=1.3.10
livekit-plugins-deepgram~=1.3.10
livekit-plugins-silero~=1.3.10
livekit-plugins-turn-detector~=1.3.10
livekit-plugins-noise-cancellation~=0.2.1
python-dotenv~=1.0
httpx~=0.28